+++ isar_flutter_libs/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.1.0
+- Update binaries to CORE/0.1.0
+
 ## 0.0.4
 - Update binaries to CORE/0.0.20