+++ isar_generator/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.1.0
+- Support for links and backlinks
+
 ## 0.0.4
 - Many improvements and fixes